Changed directory of loginscreen
